Feat/asset tools (#1)
* gui: Add tabview widget

* edit: Move asset mainctl into a new "assettool"

* gui: Fix tabview behavior when hiding tabs

* Add "New" asset tool for creating assets

* Add "Properties" asset tool

* res: Add prop_tracker

* Add WIP 3d model editor tool

* render: Preserve viewport render result in an RBO

This change reduces the invalidation of a viewport's GL widget to merely
requiring a re-blit from the RBO to the GL area.

* Add 3D view to the 3D tool

* Fix assettool crash when closing projects

* res::tracker: Fix bug when binding during acquire

This fixes a bug where on_acquire would be raised twice if the tracker
is bound to an asset during the execution of the on_asset_appear event
for that asset.

* edit::field: Add get/set for vector index

* edit::field: Fix uninitialized binding bug

* More 3D editor work

* edit::field: Fix vector index change event

* Add render highlight/marker system

This change also adds an orange highlight to the selected vertex in the
3D editing tool.

* Add vertex editing to 3D edit tool

* Fix transact issues in New asset tool

* gui::tabview: Fix warning C4099 in MSVC

    c:\projects\drnsf\src\gui.hh(1184): warning C4099: 'drnsf::gui::tabview::bar': type name first seen using 'struct' now seen using 'class' [C:\projects\drnsf\build\drnsf.vcxproj]
      c:\projects\drnsf\src\gui.hh(1138): note: see declaration of 'drnsf::gui::tabview::bar'

* gui: Separate container interface from impl

This change separates the implementation of gui::container, particularly
the widget ownership and layout propagation portion, into a separate
type gui::real_container. This allows types to implement the
gui::container interface but provide a different real_container to
contain the widgets, possibly one somewhere further up the (private)
base class hierarchy.

* gui::tabview::page: Workaround for MSVC C2886

* Move appveyor internal-test into test_script

* appveyor: Upload artifacts on test failure

* appveyor: Add .pdb files to Debug artifacts

* appveyor: Upgrade to VS2019

This change is made to possibly solve a bug with multiple static
initialization / deinitialization for inline static data member objects.

https://stackoverflow.com/questions/52809631/strange-behaviour-when-using-c17-static-inline-members-in-visual-studio

* 3d tool: Implement ctrl-clicking to pick vertices

* render::frameonly_fig: Make vertex points nicer

Vertex points are made larger, circular, and are brought slightly
closer in depth. This should also assist with clicking the points.

* gui: Add keymods to key and mousebutton events

The viewport has also been updated to use the immediate ctrl modifier
parameter rather than tracking the ctrl key status manually. This allows
ctrl-clicking to function properly when ctrl is pressed prior to the
viewport gaining focus.

* gui: Fix VK_CTRL / VK_CONTROL mixup in winapi

* 3D tool: Fix clicking on vertex 0

---------

Co-authored-by: chekwob <chek@wobbyworks.com>